A ton of 1.14.4 work Now compiles, but completely broken.

+++ src/main/java/me/desht/pneumaticcraft/client/gui/pneumatic_armor/GuiMoveStat.java
@@ -10,7 +10,7 @@ import me.desht.pneumaticcraft.client.render.pneumatic_armor.HUDHandler;
 import me.desht.pneumaticcraft.client.render.pneumatic_armor.UpgradeRenderHandlerList;
 import me.desht.pneumaticcraft.client.render.pneumatic_armor.upgrade_handler.MainHelmetHandler;
 import me.desht.pneumaticcraft.client.util.GuiUtils;
-import me.desht.pneumaticcraft.common.config.ArmorHUDLayout;
+import me.desht.pneumaticcraft.common.config.aux.ArmorHUDLayout;
 import me.desht.pneumaticcraft.common.pneumatic_armor.CommonArmorHandler;
 import net.minecraft.client.MainWindow;
 import net.minecraft.client.resources.I18n;
@@ -95,7 +95,7 @@ public class GuiMoveStat extends GuiPneumaticScreenBase {
 
     @Override
     public boolean mouseClicked(double mouseX, double mouseY, int mouseButton) {
-        if (movedStat.getBounds().contains(mouseX, mouseY)) {
+        if (movedStat.getBounds().contains((int)mouseX, (int)mouseY)) {
             if (mouseButton == 2) {
                 movedStat.setLeftSided(!movedStat.isLeftSided());
                 save();
